[snapshot] replace only specified db snapshot

+++ tools/install-db-snapshot.sh
@@ -148,20 +148,30 @@ stop_harmony() {
    return 0
 }
 
+_replace_one_db() {
+   local shard=$1
+   backupdir=$(mktemp -u "harmony_db_${shard}.backup.XXXXXX")
+   echo "backup harmony_db_${shard} to $backupdir"
+   mv -f "$HOME/harmony_db_${shard}" "$HOME/$backupdir"
+   mv -f "harmony_db_${shard}" "$HOME/"
+
+   if [ "${CONFIG_KV[cleanup]}" = "true" ]; then
+      rm -rf "${HOME:?}/$backupdir"
+      echo "cleanup: removed $backupdir"
+   fi
+}
+
 replace_db() {
-   for s in {0..3}; do
-      if [ -d "$HOME/harmony_db_${s}" ]; then
-         backupdir=$(mktemp -u harmony_db_${s}.backup.XXXXXX)
-         echo "backup harmony_db_${s} to $backupdir"
-         mv -f "$HOME/harmony_db_${s}" "$HOME/$backupdir"
-         mv harmony_db_${s} "$HOME" || return $?
-         echo "replaced harmony_db_${s}"
-         if [ "${CONFIG_KV[cleanup]}" = "true" ]; then
-            rm -rf "$backupdir"
-            echo "cleanup: removed $backupdir"
-         fi
-      fi
-   done
+   local status=0
+   case "${CONFIG_KV[shard]}" in
+      "all")
+         sid=$(find_shard_id)
+         _replace_one_db 0
+         _replace_one_db "$sid" || status=$?
+         ;;
+      0|1|2|3)
+         _replace_one_db "${CONFIG_KV[shard]}" || status=$?
+   esac
 }
 
 restart_harmony() {
